% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train_model_music.R
\name{add_deconvolution_training_model_music}
\alias{add_deconvolution_training_model_music}
\title{add_deconvolution_training_model_music}
\usage{
add_deconvolution_training_model_music(
    transcriptome_data,
    model_name,
    subtype_vector
)
}
\arguments{
\item{transcriptome_data}{Path to transcriptomic data to be
used for training. Has to contain the cell subtypes to which the
similarity has to be calculated. Note that the first column has
to contain the HGNC symbols and the header not! not the first
sample name but a mere label for this HGNC row.}

\item{model_name}{Name of the model}

\item{subtype_vector}{Character vector containing the subtype
labels of the training data samples}
}
\value{
Stores a new model in the package directory
}
\description{
\code{add_deconvolution_training_model_music} adds a new model
}
\examples{
data("Lawlor") # Data from Lawlor et al.
data(meta_data)

subtype_vector = meta_data$Subtype # extract the training sample subtype labels
add_deconvolution_training_model_music(
    transcriptome_data = Lawlor,
    model_name = "my_model",
    subtype_vector = subtype_vector
)
}
