% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{configure_graphics}
\alias{configure_graphics}
\title{configure_graphics}
\usage{
configure_graphics()
}
\value{
Heatmap related graphics configuration
}
\description{
\code{configure_graphics} configure the graphics paramters
}
\examples{
graphical_configuration = configure_graphics()
}
