% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{create_heatmap_deconvolution}
\alias{create_heatmap_deconvolution}
\title{create_heatmap_deconvolution}
\usage{
create_heatmap_deconvolution(
    visualization_data,
    deconvolution_results,
    aggregate_differentiated_stages,
    confidence_threshold,
    show_colnames,
    Graphics_parameters,
    high_threshold,
    low_threshold,
    utilize_sadanandam_genes
)
}
\arguments{
\item{visualization_data}{Matrix of the transcriptome data
that shall be visualized. Notice the convention
that the row names have to contain the HGNC identifier.}

\item{deconvolution_results}{The dataframe returned
by the deconvolution analysis.}

\item{aggregate_differentiated_stages}{Show the differentiation stage similarities
aggregated over all differentiated stages - alpha, beta, gamma, delta
accinar and ductal - or specific for each differentiation stage.
Default value FALSE, alternative value TRUE.}

\item{confidence_threshold}{Threshold above which deconvolutions are deemed unsuccessful
and corresponding results being masked on the the plots. Default value 1.1.}

\item{show_colnames}{Whether to show the sample column names. Default value FALSE.}

\item{Graphics_parameters}{Pheatmap visualization paramters.
You can customize visualization colors.
Read the vignette for more information.}

\item{high_threshold}{Threshold depending on which a deconvolution result
is interpreted as 'high'. If not set, a statistical estimation will approximately
identify a signficance threshold for a high similarity.}

\item{low_threshold}{Threshold depending on which a deconvolution result
is interpreted as 'low'. If not set, a statistical estimation will approximately
identify a signficance threshold for a low similarity.}

\item{utilize_sadanandam_genes}{Whether to utilize the same genes as the 
Sadanandam et al publication which have been utilize to create the DECO
manuscript visualizations. Default value FALSE.}
}
\value{
Plots
}
\description{
\code{create_heatmap_deconvolution}
visualizes the deconvolution results as heatmap.
Please note that the rows of the expression data matrix 
are expected to be HGNC symbols and columns are expected 
to contain the samples.
}
\examples{
data(deconvolution_results, envir = environment())
data(visualization_data, envir = environment())

create_heatmap_deconvolution(
    visualization_data = visualization_data,
    deconvolution_results = deconvolution_results,
    aggregate_differentiated_stages = FALSE,
    confidence_threshold = 1.1,
    show_colnames = FALSE,
    Graphics_parameters = "",
    high_threshold = 66,
    low_threshold = 33,
    utilize_sadanandam_genes = FALSE
)
}
