% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{create_PCA_differentiation_stages}
\alias{create_PCA_differentiation_stages}
\title{create_PCA_differentiation_stages}
\usage{
create_PCA_differentiation_stages(
    transcriptome_file_path,
    deconvolution_results,
    annotation_columns,
    Graphics_parameters,
    baseline_mode
)
}
\arguments{
\item{transcriptome_file_path}{Path to the transcriptome
data that shall be visualized. Notice the convention
that the first row has to contain the HGNC identifier}

\item{deconvolution_results}{The dataframe returned
by the deconvolution analysis}

\item{annotation_columns}{Column names that are to be visualized
ontop of the correlation matrix}

\item{Graphics_parameters}{Pheatmap visualization paramters.
You can customize visualization colors.
Read the vignette for more information.}

\item{baseline_mode}{Which measurement represents the baseline
of the differentiation similarity: 'absolute' = maximal
similarity of the training sample to its subtype e.g.
alpha cell similarity to alpha cell =}
}
\value{
Plots
}
\description{
\code{create_PCA_differentiation_stages}
visualizes the differentiation stage predictions as PCA.
Please note that the first column of the expression
data matrix has to contain the HGNC identifier
}
\examples{
meta_data_path = system.file(
    "Data/Meta_information/Meta_information.tsv",
    package = "artdeco"
)
meta_data      = read.table(
    meta_data_path, sep ="\\t",
    header = TRUE,
    stringsAsFactors = FALSE
)
rownames(meta_data) = meta_data$Sample

transcriptome_file_path = system.file(
    "/Data/Expression_data/Visualization_PANnen.tsv",
     package = "artdeco"
)
create_PCA_differentiation_stages(
    transcriptome_file_path = transcriptome_file_path,
    deconvolution_results = meta_data,
    annotation_columns = c(
        "Differentiation_Stages_Subtypes",
        "Differentiation_Stages_Aggregated",
        "Differentiatedness"
     ),
     Graphics_parameters = "",
     baseline_mode = "absolute"
)
}
