% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolve_bseq_sc.R
\name{Deconvolve_bseq_sc}
\alias{Deconvolve_bseq_sc}
\title{Deconvolve_bseq_sc}
\usage{
Deconvolve_bseq_sc(deconvolution_data, models_list, models, nr_permutations)
}
\arguments{
\item{deconvolution_data}{Data to be deconvolved.}

\item{models_list}{List of models.}

\item{models}{Models utilized.}

\item{nr_permutations}{Amount of permutations.}
}
\value{
Matrix containing the deconvolution results.
}
\description{
\code{Deconvolve_bseq_sc} Utilizes BSeq-sc for deconvolution. Wrapper function.
}
