% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolve_music.R
\name{Deconvolve_music}
\alias{Deconvolve_music}
\title{Deconvolve_music}
\usage{
Deconvolve_music(deconvolution_data, models_list, models, nr_permutations)
}
\arguments{
\item{deconvolution_data}{A data frame that contains the gene expression 
data. Rows are expected to be HGNC symbols and columns are expected to 
contain the samples.}

\item{models_list}{List of models to be used.}

\item{models}{Which model to use.}

\item{nr_permutations}{Amount perturbations.}
}
\value{
Matrix containing the deconvolution results.
}
\description{
\code{Deconvolve_music} Utilizes MuSiC for deconvolution. Wrapper function.
}
