% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolve_NMF.R
\name{Deconvolve_NMF}
\alias{Deconvolve_NMF}
\title{Deconvolve_NMF}
\usage{
Deconvolve_NMF(deconvolution_data, models_list, models, nr_permutations)
}
\arguments{
\item{deconvolution_data}{Data to be deconvolved}

\item{models_list}{List of models}

\item{models}{Models utilized}

\item{nr_permutations}{Amount of permutations}
}
\description{
Deconvolve_NMF
}
