% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolve_transcriptome.R
\name{Deconvolve_transcriptome}
\alias{Deconvolve_transcriptome}
\title{Deconvolve_transcriptome}
\usage{
Deconvolve_transcriptome(
    transcriptome_data,
    deconvolution_algorithm,
    models,
    nr_permutations,
    output_file
)
}
\arguments{
\item{transcriptome_data}{A data frame that contains the gene 
expression data. Rows are expected to be HGNC symbols and columns are 
expected to contain the samples.}

\item{deconvolution_algorithm}{Which deconvolution algorithm to choose
from. Options: 'music','bseqsc' (CIBERSORT), 'nmf'.}

\item{models}{List of models to be used. Use show_models_NMF(),
show_models_music() or show_models_bseqsc()
to view available models or add new model via
add_deconvolution_training_model_*().}

\item{nr_permutations}{Utilized to calculate p-value
Higher amount of permutations generally lead to more
precise p-value estimates. Default value 1000.}

\item{output_file}{Path of output file. If not specified,
no hard-disk written output will occur.}
}
\value{
Similarity measurements of differentiation stages.

Please note when using the \code{music} or \code{bseqsc} algorithm, that
you need to load the corresponding libraries first. See the Wiki of the 
artdeco Github page for more information.
}
\description{
\code{Deconvolve_transcriptome} measures the similarity
of one or many query RNA-seq or microarray samples to samples
with known differentiation stage contained in the training models.
}
\examples{
data("visualization_data")
Deconvolve_transcriptome(
    transcriptome_data = visualization_data
)
}
