% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Determine_differentiation_stage.R
\name{Determine_differentiation_stage}
\alias{Determine_differentiation_stage}
\title{Determine_differentiation_stage}
\usage{
Determine_differentiation_stage(
    transcriptome_file_path,
    models,
    nr_permutations,
    p_value,
    meta_data,
    output_file
)
}
\arguments{
\item{transcriptome_file_path}{Path to the file which contains
the transcriptome data. Notice the HGNC row convention:
Rownames have to include the unique HGNC identifier, see vignette.}

\item{models}{List of models to be used. Use show_models()
to view available models or add new model via
add_deconvolution_training_model()}

\item{nr_permutations}{Utilized to calculate p-value
Higher amount of permutations generally lead to more
precise p-value estimates}

\item{p_value}{P-value that determines when a deconvolution
is to be considered significant. Lower p-values lead to loss
of sensitivity but increase in specificity.}

\item{meta_data}{Dataframe that stores the meta information
of the transcriptomes. Can be build by the function itself or
handed over}

\item{output_file}{Path of output file. If not specified,
no hard-disk written output will occur.}
}
\value{
Similarity measurements to differentiation
stages
}
\description{
\code{Determine_differentiation_stage} measures the similarity
of one or many query RNA-seq or microarray samples to samples
with known differentiation stage contained in the training models.
}
\examples{
transcriptome_file_path = system.file(
"/Data/Expression_data/PANnen_Test_Data.tsv", package = "artdeco")
Determine_differentiation_stage(
    transcriptome_file_path = transcriptome_file_path,
    models = "Alpha_Beta_Gamma_Delta_Lawlor"
)
}
