% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{prepare_result_matrix}
\alias{prepare_result_matrix}
\title{prepare_result_matrix}
\usage{
prepare_result_matrix(
    prediction_stats_list,
    parameter_list,
    p_value,
    meta_data,
    models
)
}
\arguments{
\item{prediction_stats_list}{list containing the fitting results}

\item{parameter_list}{List containing the model parameters}

\item{p_value}{P-value significance threshold}

\item{meta_data}{Contains the deconvolution results as dataframe}

\item{models}{List of training models}
}
\value{
list of models
}
\description{
Prepares the deconvolution results
}
