% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{quantify_similarity}
\alias{quantify_similarity}
\title{quantify_similarity}
\usage{
quantify_similarity(
    meta_data,
    subtypes,
    model,
    fit,
    parameter_list,
    baseline,
    not_sig_samples
)
}
\arguments{
\item{meta_data}{deconvolution result dataframe}

\item{subtypes}{list of cell subtypes}

\item{model}{Deconvolution model}

\item{fit}{Deconvolution fit}

\item{parameter_list}{List of parameters of the models}

\item{baseline}{Absolut or relative mode}

\item{not_sig_samples}{list of not significantly deconvolable samples}
}
\value{
dataframe containing baseline adjusted similarities
}
\description{
\code{quantify_similarity} custom function that allows to
alternate between different baselines for similarity
}
