% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{remove_model}
\alias{remove_model}
\title{remove_model}
\usage{
remove_model(
    model_name,
    test_mode
)
}
\arguments{
\item{model_name}{Name of the model}

\item{test_mode}{Testrun indicator}
}
\value{
list of models
}
\description{
\code{remove_model} removes a model
}
\examples{
model = list.files(system.file("Models/", package = "artdeco"))[1]
remove_model(
    model_name = model,
    test_mode = TRUE
)
}
