% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{remove_model_bseqsc}
\alias{remove_model_bseqsc}
\title{remove_model_bseqsc}
\usage{
remove_model_bseqsc(
    model_name,
    test_mode
)
}
\arguments{
\item{model_name}{Name of the model.}

\item{test_mode}{Testrun indicator. Default value FALSE.}
}
\value{
Removal of model from package directory
}
\description{
\code{remove_model_bseqsc} removes a model from the bseqsc library
}
\examples{
remove_model_bseqsc(
    model_name = "My_model",
    test_mode = TRUE
)
}
