% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{remove_model_music}
\alias{remove_model_music}
\title{remove_model_music}
\usage{
remove_model_music(
    model_name,
    test_mode
)
}
\arguments{
\item{model_name}{Name of the model}

\item{test_mode}{Testrun indicator. Default value FALSE.}
}
\description{
\code{remove_model_music} removes a model from the music library
}
\examples{
remove_model_music(
    model_name = "My_model",
    test_mode = TRUE
)
}
