% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{show_models}
\alias{show_models}
\title{show_models}
\usage{
show_models(
lib_name
)
}
\arguments{
\item{lib_name}{Specify the library (e.g. "NMF", "music", "bseqsc" or "all").}
}
\value{
list of models
}
\description{
\code{show_models} shows which models have been trained for what library
}
\examples{
show_models(
lib_name = "NMF"
)
}
