% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{show_models_bseqsc}
\alias{show_models_bseqsc}
\title{show_models_bseqsc}
\usage{
show_models_bseqsc()
}
\value{
list of models
}
\description{
\code{show_models_bseqsc} shows which models have been trained for bseqsc
}
\examples{
show_models_bseqsc()
}
