% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{show_models_music}
\alias{show_models_music}
\title{show_models_music}
\usage{
show_models_music()
}
\value{
list of models
}
\description{
\code{show_models_music} shows which models have been trained
}
\examples{
show_models_music()
}
