% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{show_models_NMF}
\alias{show_models_NMF}
\title{show_models_NMF}
\usage{
show_models_NMF()
}
\value{
list of models
}
\description{
\code{show_models_NMF} shows which models have been trained for 
the NMF method
}
\examples{
show_models_NMF()
}
