% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{simulateCounts}
\alias{simulateCounts}
\title{Simulate count data based on negative binomial distribution}
\usage{
simulateCounts(simParam, nSamples, nGenes, simSeed = NULL, simMarkers = TRUE)
}
\arguments{
\item{simParam}{List of parameters estimated by
\code{\link{estimateParameters}}.}

\item{nSamples}{A integer specifing the number of samples to be simulated.}

\item{nGenes}{A integer specifing the number of genes to be simulated.}

\item{simSeed}{A integer specifing seed for simulation}

\item{simMarkers}{Logical, indicating whether genes to be simulated are
markers or not}
}
\value{
Matrix with simulated count values.
}
\description{
\code{simulateCounts} simulates the expression values using
a negative binomial distribution with parameters estimated by
\code{\link{estimateParameters}}. This function is used by
\code{\link{simulateCellTypes}}.
}
