% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{simulateNegativeControls}
\alias{simulateNegativeControls}
\title{Simulate expression data which does not follow a cell-type specific profile.}
\usage{
simulateNegativeControls(
    nGenes,
    numSamples,
    normMean,
    normSD,
    seed,
    verbose
)
}
\arguments{
\item{nGenes}{A integer specifing the number of genes to be simulated.}

\item{numSamples}{A integer specifing the number of samples to be simulated.}

\item{normMean}{A integer specifing the mean parameter of the normal distribution
which is used to generate mean expression values for the simulation.}

\item{normSD}{A integer specifing the standard deviation parameter of the
normal distribution which is used to to generate mean expression values for
the simulation.}

\item{seed}{A integer specifing seed for simulation. Default is \code{NULL}.
Random seed will be generated if no seed is provided.}

\item{verbose}{Logical, indicating whether status updates will be printed.
Default is \code{TRUE}.}
}
\value{
Matrix with simulated count values. Samples as columns and genes
 as rows.
}
\description{
\code{simulateNegativeControls} simulates random expression profiles
following a negative binomial distribution. Parameters of the negative
binomial are randomly sampled from a normal distribution.
Can be used as negative controls for benchmarking purposes.
}
