% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train_model_bseqsc.R
\name{add_deconvolution_training_model_bseqsc}
\alias{add_deconvolution_training_model_bseqsc}
\title{add_deconvolution_training_model_bseqsc}
\usage{
add_deconvolution_training_model_bseqsc(
    transcriptome_data,
    model_name,
    subtype_vector,
    marker_gene_list,
    training_p_value_threshold,
    training_nr_permutations,
    training_nr_marker_genes
)
}
\arguments{
\item{transcriptome_data}{Path to transcriptomic data to be
used for training. Has to contain the cell subtypes to which the
similarity has to be calculated. Note that the first column has
to contain the HGNC symbols and the header not! not the first
sample name but a mere label for this HGNC row.}

\item{model_name}{Name of the model}

\item{subtype_vector}{Character vector containing the subtype
labels of the training data samples}

\item{marker_gene_list}{List the contains the marker genes
for each subtype. Has to be in the type of list() with each
subtype being an entry.}

\item{training_p_value_threshold}{P-value at which a training
is deemed successfull}

\item{training_nr_permutations}{Amount of perturbation which
result in a p-value. Higher number of perturbation generally
improves the p-value estiamtes}

\item{training_nr_marker_genes}{How many genes should be utilized
as list of marker genes}
}
\value{
Stores a new model in the package directory
}
\description{
\code{add_deconvolution_training_model_bseqsc} adds a new model
}
\examples{
data("Lawlor") # Data from Lawlor et al.
data(meta_data)

subtype_vector = as.character(meta_data$Subtype) # extract the training sample subtype labels

add_deconvolution_training_model_bseqsc(
    transcriptome_data = Lawlor,
    model_name = "my_model",
    subtype_vector = subtype_vector,
    training_nr_permutations = 10,
    training_nr_marker_genes = 100
)
}
