% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train_model_music.R
\name{add_deconvolution_training_model_music}
\alias{add_deconvolution_training_model_music}
\title{add_deconvolution_training_model_music}
\usage{
add_deconvolution_training_model_music(
    transcriptome_data,
    model_name,
    subtype_vector
)
}
\arguments{
\item{transcriptome_data}{Path to transcriptomic data to be
used for training. Has to contain the cell subtypes to which the
similarity will be calculated. Note that the row names have
to contain the HGNC symbols and the column names the sample names.}

\item{model_name}{Name of the model.}

\item{subtype_vector}{Character vector containing the subtype
labels of the training data samples (\code{transcriptome_data}).}
}
\value{
Stores a new model in the package directory

Please note when using this function, that you need to load the 
corresponding library first.
}
\description{
\code{add_deconvolution_training_model_music} adds a new model
}
\examples{
library("MuSiC")
library("xbioc")

data("Lawlor") # Data from Lawlor et al.
data(meta_data)

# extract the training sample subtype labels
subtype_vector = as.character(meta_data$Subtype) 
add_deconvolution_training_model_music(
    transcriptome_data = Lawlor,
    model_name = "my_model",
    subtype_vector = subtype_vector
)
}
