% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{create_PCA_deconvolution}
\alias{create_PCA_deconvolution}
\title{create_PCA_deconvolution}
\usage{
create_PCA_deconvolution(
   visualization_data,
   deconvolution_results,
   aggregate_differentiated_stages,
   confidence_threshold,
   Graphics_parameters,
   high_threshold,
   low_threshold,
   utilize_sadanandam_genes
)
}
\arguments{
\item{visualization_data}{Transcriptome data that shall be visualized.
Notice the convention that the row names have to contain the HGNC identifier}

\item{deconvolution_results}{The dataframe returned
by the deconvolution analysis}

\item{aggregate_differentiated_stages}{Whether to aggregate the deconvolution
results. Default vale FALSE.}

\item{confidence_threshold}{Threshold at which to discard a prediction. 
Default value 1.1}

\item{Graphics_parameters}{Pheatmap visualization paramters.
You can customize visualization colors.
Read the vignette for more information.}

\item{high_threshold}{Threshold at which to define a deconvolution results
as 'high'. Default value 101.}

\item{low_threshold}{Threshold at which to define a deconvolution results
as 'low'. Default value 0.}

\item{utilize_sadanandam_genes}{Whether to utilize the same genes as the 
Sadanandam et al publication which have been utilize to create the DECO
manuscript visualizations. Default value FALSE.}
}
\description{
\code{create_PCA_deconvolution}
visualizes the differentiation stage predictions as PCA.
Please note that the rows of the expression data matrix 
are expected to be HGNC symbols and columns are expected 
to contain the samples.
}
\examples{
data(deconvolution_results, envir = environment())
data(visualization_data, envir = environment())
create_PCA_deconvolution(
    visualization_data = visualization_data,
    deconvolution_results = deconvolution_results,
    Graphics_parameters = "",
    utilize_sadanandam_genes = TRUE
)
}
