% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_marker_genes.R
\name{identify_marker_genes}
\alias{identify_marker_genes}
\title{identify_marker_genes}
\usage{
identify_marker_genes(
  expression_training_mat,
  subtype_vector,
  subtype,
  training_nr_marker_genes
)
}
\arguments{
\item{expression_training_mat}{matrix that contains the expression data}

\item{subtype_vector}{Vector containing the subtype of the training samples}

\item{subtype}{Subtype for which to determine the differentially expressed}

\item{training_nr_marker_genes}{Number of desired marker genes to describe a subtype}
}
\value{
Returns marker genes
}
\description{
identify_marker_genes
}
