% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{remove_model}
\alias{remove_model}
\title{remove_model}
\usage{
remove_model(
    model_name,
    lib_name,
    test_mode
)
}
\arguments{
\item{model_name}{Name of the model}

\item{lib_name}{Name of the library that contains the model (e.g. "NMF", "music" or "bseqsc")}

\item{test_mode}{Testrun indicator. Default value FALSE.}
}
\description{
\code{remove_model} removes a model from a library
}
\examples{
remove_model(
    model_name = "My_model",
    lib_name = "bseqsc",
    test_mode = TRUE
)
}
