% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{remove_model_NMF}
\alias{remove_model_NMF}
\title{remove_model_NMF}
\usage{
remove_model_NMF(
    model_name,
    test_mode
)
}
\arguments{
\item{model_name}{Name of the model}

\item{test_mode}{Testrun indicator}
}
\description{
\code{remove_model_NMF} removes a model from the NMF library
}
\examples{
remove_model_NMF(
    model_name = "My_model",
    test_mode = TRUE
)
}
